backend/requirements.txt
django==1.11.8
djangorestframework
djangorestframework-jwt
psycopg2-binary
dj-database-url